requirements.txt
Flask==3.1.2
flask-cors
python-dotenv==1.1.0
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-google-genai==2.1.12
langchain-pinecone==0.2.8
sentence-transformers==4.1.0
pypdf==5.6.0
crewai
crewai-tools
-e .
